Modularize the program ABI and host surface
